+++ src/components/create-reaview/CreateReviewForm.tsx
@@ -1,7 +1,5 @@
 "use client";
 
-import { useMutation } from "@tanstack/react-query";
-import { useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import { FormProvider, useForm } from "react-hook-form";
 import SolidButton from "@/components/common/buttons/SolidButton";
@@ -9,9 +7,7 @@ import CommonTextArea from "@/components/common/inputs/TextArea";
 import RatingInput from "@/components/create-reaview/RatingInput";
 import ReviewImageInput from "@/components/create-reaview/ReviewImageInput";
 import useCookie from "@/hooks/auths/useTokenState";
-
-import { fetcher } from "@/lib/user/fetcher";
-import useUserStore from "@/store/auth/useUserStore";
+import { useReviewMutation } from "@/hooks/review/useReviewMutation";
 
 interface ReviewFormData {
   rating: number;
@@ -22,9 +18,7 @@ interface ReviewFormData {
 
 export default function CreateReviewForm({ meetupId }: { meetupId: string }) {
   const token = useCookie("accessToken");
-  const { user } = useUserStore();
   const router = useRouter();
-  const queryClient = useQueryClient();
 
   if (!meetupId) {
     router.replace("/");
@@ -47,33 +41,7 @@ export default function CreateReviewForm({ meetupId }: { meetupId: string }) {
   const content = watch("content");
   const isFormValid = rating !== -1 && content.trim() !== "";
 
-  const createReviewMutation = useMutation({
-    mutationFn: async (formData: FormData) => {
-      if (!token) throw new Error("로그인이 필요합니다.");
-
-      const res = await fetcher("/reviews", token, {
-        method: "POST",
-        body: formData,
-        auth: true,
-      });
-
-      if (!res.ok) {
-        const errorData = await res.json().catch(() => null);
-        console.error("서버 응답 에러:", errorData);
-        throw new Error("리뷰 작성에 실패했습니다.");
-      }
-      return res.json();
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: ["meetings", "myReview", "study"],
-      });
-      queryClient.invalidateQueries({
-        queryKey: ["meetings", "myReview", "tutoring"],
-      });
-      router.replace(`/user/${user?.userId}`);
-    },
-  });
+  const createReviewMutation = useReviewMutation();
 
   const handleRatingChange = (value: number) => {
     setValue("rating", value);
@@ -84,6 +52,8 @@ export default function CreateReviewForm({ meetupId }: { meetupId: string }) {
   };
 
   const onSubmit = handleSubmit(async (data) => {
+    if (!token) return;
+
     const formData = new FormData();
 
     const requestData = {
@@ -105,7 +75,7 @@ export default function CreateReviewForm({ meetupId }: { meetupId: string }) {
       formData.append("image", new Blob(), "");
     }
 
-    await createReviewMutation.mutateAsync(formData);
+    await createReviewMutation.mutateAsync({ formData, token });
   });
 
   return (
